Add Snowflake as db output option
I have mostly just copied the setup for MySQL and adjusted it to work
with Snowflake. I also changed the name and the CLI because it felt
stupid to run `dtd2mysql --snowflake` or something like that lol

The tests are all passing, but I have not tried out actually loading
into snowflake yet.

Please refer to the README for more details.